+++ projects/label-designer/docs/CHANGELOG.md
@@ -1,3 +1,14 @@
+<a name="3.3.0"></a>
+# 3.3.0 (2020-02-??)
+
+### Features
+* **editor:** added option to skip n labels from the start of the pdf.
+* **editor:** added option to repeat labels n times when generating the pdf.
+* **editor:** added option to include separator only if following field is empty.
+
+### Bug fix
+* **core:** remove empty page between pages  
+
 <a name="3.2.4"></a>
 # 3.2.4 (2019-12-20)